+++ packages/jest-config/jest.config.js
@@ -1,0 +1,3 @@
+import jestConfig from '../../settings/jest.config.js';
+
+export default jestConfig;